switch restart always to unless-stopped
I try to integrate the backup mechanisms of https://github.com/blacklabelops/volumerize, handling named volumes and automatic service stop and restart for consistent backups. Nevertheless this needs a clean stop of a service, which is only able to be achieved, if the docker restarts the containers only if they are not explicitly stopped.