hotfix tagging of webapp builds
since the introduction of the php base image version information in the webapp container are doubled. by removing the build args from the php image at least there is no "newest" version any longer for webapp, but it still shows the version twice.

this introduces a hotfix for this behaviour.

Signed-off-by: Felix Bartels <felix@host-consultants.de>