Skip downloading of packages if folder already exists
The base image is using onbuild instructions to reduce code duplication (in regards to fetching Kopano packages). Sadly the unbuild instructions are not only executed on the directly following container image, but in every image that is using the images from this project as a base.

Fixes https://github.com/zokradonh/kopano-docker/issues/430

Signed-off-by: Felix Bartels <felix@host-consultants.de>